restaurant_backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
mysqlclient==2.2.4
pika~=1.3.2
requests~=2.32.3

